Move location of pre-existing test_file_selection
